Propose download_sources.sh usage as an alternative to the wget command line.
I've ported the download_sources.sh script to work more flexibly and proper with the CMLFS. Basically, i've added the possibility to download sources where you'd want, but with /mnt/cmlfs/sources/pkgs as default.

I didn't too much in this weekend because i've needed to study to finals (i've had physics yesterday and i couldn't lose the chance to get a good grade in it).

May i'll "convert" the plain-text/markdown manual to HTML at the end of this week (and of the finals).